+++ src/pytest_inmanta_lsm/plugin.py
@@ -9,7 +9,7 @@
 import logging
 import os
 import os.path
-from typing import Iterator
+from typing import Dict, Iterator, Union
 
 import pytest
 from pytest_inmanta.plugin import Project
@@ -70,7 +70,20 @@ def get_opt_or_env_or(config, key: str, default: str) -> str:
 
 
 @pytest.fixture
-def remote_orchestrator(project: Project, request) -> "Iterator[r_orchestrator.RemoteOrchestrator]":
+def remote_orchestrator_settings() -> Dict[str, Union[str, int, bool]]:
+    """Override this fixture in your tests or conf test to set custom environment settings after cleanup. The supported
+    settings are documented in https://docs.inmanta.com/inmanta-service-orchestrator/3/reference/environmentsettings.html
+
+    The remote_orchestrator fixture already sets a number of non-default values to make the fixture work as it should.
+    However, overriding for example the deploy interval so speed up skip resources can be useful.
+    """
+    return {}
+
+
+@pytest.fixture
+def remote_orchestrator(
+    project: Project, request, remote_orchestrator_settings
+) -> "Iterator[r_orchestrator.RemoteOrchestrator]":
     LOGGER.info("Setting up remote orchestrator")
 
     env = get_opt_or_env_or(request.config, "inm_lsm_env", "719c7ad5-6657-444b-b536-a27174cb7498")
@@ -78,7 +91,20 @@ def remote_orchestrator(project: Project, request) -> "Iterator[r_orchestrator.R
     user = get_opt_or_env_or(request.config, "inm_lsm_remote_user", "centos")
     noclean = get_opt_or_env_or(request.config, "inm_lsm_noclean", "false").lower() == "true"
 
-    remote_orchestrator = r_orchestrator.RemoteOrchestrator(host, user, env, project)
+    # set the defaults here and lets the fixture override specific values
+    settings = {
+        "auto_deploy": True,
+        "server_compile": True,
+        "agent_trigger_method_on_auto_deploy": "push_incremental_deploy",
+        "push_on_auto_deploy": True,
+        "autostart_agent_deploy_splay_time": 0,
+        "autostart_agent_deploy_interval": 600,
+        "autostart_agent_repair_splay_time": 600,
+        "autostart_agent_repair_interval": 0,
+    }
+    settings.update(remote_orchestrator_settings)
+
+    remote_orchestrator = r_orchestrator.RemoteOrchestrator(host, user, env, project, settings, noclean)
     remote_orchestrator.clean()
 
     yield remote_orchestrator